Move upgrade step in update_250 and create specific update step entry
