+++ src/Products/urban/migration/update_250.py
@@ -9,7 +9,7 @@ from plone import api
 from plone.app.textfield import RichTextValue
 from plone.app.uuid.utils import uuidToObject
 
-from Products.urban.interfaces import IGenericLicence
+from Products.urban.interfaces import IGenericLicence, IBaseBuildLicence
 import logging
 import re
 
@@ -314,5 +314,30 @@ def migrate_add_tax_other_option(context):
     for licence_config in licence_configs:
         if "other" not in licence_config.tax:
             licence_config.tax.invokeFactory('UrbanVocabularyTerm', id='other', title="Autre")
+
+    logger.info("migration step done!")
+def migrate_move_basebuildlicence_architects_and_geometricians_to_representative_contacts(context):
+    """
+    """
+    logger = logging.getLogger('urban: migrate migrate_move_basebuildlicence_architects_and_geometricians_to_representative_contacts')
+    logger.info("starting migration step")
+    catalog = api.portal.get_tool('portal_catalog')
+    licence_brains = catalog(object_provides=IBaseBuildLicence.__identifier__)
+    licences = [li.getObject() for li in licence_brains]
+    for licence in licences:
+        architects = licence.getField('architects')
+        if architects:
+            for architect in architects.get(licence):
+                rc_list = licence.getRepresentativeContacts()
+                rc_list.append(architect)
+                licence.setRepresentativeContacts(rc_list)
+                licence.setArchitects([])
+        geometricians = licence.getField('geometricians')
+        if geometricians:
+            for geometrician in geometricians.get(licence):
+                rc_list = licence.getRepresentativeContacts()
+                rc_list.append(geometrician)
+                licence.setRepresentativeContacts(rc_list)
+                licence.setGeometricians([])
 
     logger.info("migration step done!")
