+++ src/Products/urban/migration/upgrades.zcml
@@ -275,13 +275,20 @@
         handler=".update_250.migrate_notaryletter_specificfeatures_texts"
         profile="Products.urban:default" />
 
-
     <gs:upgradeStep
         title="Add Tax vocabulary 'Other' option"
         description="urban upgrade step"
         source="1105"
         destination="1106"
         handler=".update_250.migrate_add_tax_other_option"
+        profile="Products.urban:default" />
+
+    <gs:upgradeStep
+        title="Move architects and geometricians to representative contacts for base buildlicence"
+        description="urban upgrade step"
+        source="1106"
+        destination="1107"
+        handler=".update_250.migrate_move_basebuildlicence_architects_and_geometricians_to_representative_contacts"
         profile="Products.urban:default" />
 
 </configure>
